Added glad + copyright labels
- Added glad directly into the repo to avoid package manager messiness
- Added very very basic OpenGL glad + glfw test window
- Updated README.md to include info for Fedora/RHEL-based Distros
- Fixed label enum conversion and moved into own folder (for future translation stuff)
- Some comments on the Biome file
- Removed non-existent animations
- Added copyright labels for files I created (missing those from jcbbc)
- Some TODOs added
- Moved animation, packets and sounds to "network" enums
- Renamed animations to animation_ids
- Renamed Packets to packet_ids
- renamed sounds to world_event_ids

+++ src/bpp_shared/strings/labels.cpp
@@ -1,13 +1,21 @@
+/*
+ * Copyright (c) 2026, Pixel Brush <pixelbrush.dev>
+ *
+ * SPDX-License-Identifier: GPL-3.0-only
+ * 
+*/
+
 #include "labels.h"
+#include <cstddef>
 
 // Get the appropriate Label for the passed id
 std::string IdToLabel(int16_t id) {
 	if (id > SLOT_EMPTY) {
 		if (id < BLOCK_MAX) {
-			return blockLabels[id];
+			return blockLabels[size_t(id)];
 		}
 		if (id >= ITEM_SHOVEL_IRON) {
-			return itemLabels[id - ITEM_SHOVEL_IRON];
+			return itemLabels[size_t(id - ITEM_SHOVEL_IRON)];
 		}
 	}
 	return "Invalid";
